new folder structure for extractions #118
