+++ scripts/extractions/patch_extractions/extract_patch.py
@@ -11,7 +11,6 @@ from pathlib import Path
 import geopandas as gpd
 import pandas as pd
 import requests
-from extract_common import generate_output_path, pipeline_log, post_job_action
 from extract_meteo import create_datacube_meteo, create_job_dataframe_meteo
 from extract_optical import create_datacube_optical, create_job_dataframe_s2
 from openeo.rest import OpenEoApiError, OpenEoApiPlainError, OpenEoRestError
@@ -20,6 +19,11 @@ from openeo_gfmap.backend import cdse_connection
 from openeo_gfmap.manager.job_manager import GFMAPJobManager
 from openeo_gfmap.manager.job_splitters import load_s2_grid, split_job_s2grid
 
+from worldcereal.openeo.extract_common import (
+    generate_output_path,
+    pipeline_log,
+    post_job_action,
+)
 from worldcereal.stac.constants import ExtractionCollection
 
 from extract_sar import (  # isort: skip
