+++ scripts/extractions/point_extractions/extract_point.py
